getCommunitydata fallback function
When the community data is unavailable in the community state which will happen if the user navigates to the post page directly as the community data will be stored in state. Which in our case only populates via SSR on community page render.
This fallback function monitors the state when we make a request with community data but such data doesn't yet exist in state, it will fetch the community data and store it in state so it can now be used across the session.